proxy_top_restructure
Corrected a serialization error by moving the  proxy scoring function to a serializable class.
